Add files via upload
Added the 2-opt method and the swap places method required for the 2-opt. Also, added the taxicab metric to calculate distance between 2 nodes (which is a cool way of saying i am not calculating the hypotenuse i am summing the other two sides of the triangle.). Added the read from file method which was previously written in test to NodeList. Also, the distance calculating method forgot to empty the shopping-card so it is fixed too. THE FILEREADER AND THE CALCULATETWOOPT METHOD REQUIRES FIXING THE REASON IS EXPLAINED IN THE CODE AS COMMENTS.